Refactor CompoundSelector to separate type.
